signsToJson/requirements.txt
# Core dependencies
PyMuPDF==1.24.0
Pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0
python-dotenv==1.0.1

# CLI
click==8.1.7
